Chores has been updated to new standards!

+++ Services/ChoresService.cs
@@ -11,18 +11,41 @@ public class ChoresService
 
     internal Chore Create(Chore choreData)
     {
-        Chore newChore = _repo.Create(choreData);
-        return newChore;
+        Chore chore = _repo.Create(choreData);
+        return chore;
     }
-    internal List<Chore> Get()
+
+    internal List<Chore> Get(string userId)
     {
         List<Chore> chores = _repo.Get();
         return chores;
     }
 
-    internal string Remove(int id)
+    internal Chore GetOne(int id, string userId)
     {
-        string message = _repo.Remove(id);
-        return message;
+        Chore chore = _repo.GetOne(id);
+        if (chore == null)
+        {
+            throw new Exception("No chore at that id!");
+        }
+        if (chore.CreatorId != userId)
+        {
+            throw new Exception("hey! Not yours! The ID does not match!");
+        }
+        return chore;
+    }
+
+
+
+    internal string Remove(int id, string userId)
+    {
+        Chore original = GetOne(id, userId);
+        if (original.CreatorId != userId)
+        {
+            throw new Exception("Yo. Not yours buddy.");
+        }
+        //NOTE CHECK postItSharp if you'd like to see a soft delete!!
+        _repo.Remove(id);
+        return $"{original.Name} has turned to DUST.";
     }
 }
